striteri y strmapi cambiadas i a unsigned int y tests striteri

+++ 01_libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: jrollon- <jrollon-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/20 20:53:52 by jrollon-          #+#    #+#             */
-/*   Updated: 2025/01/20 21:00:45 by jrollon-         ###   ########.fr       */
+/*   Updated: 2025/01/24 20:28:43 by jrollon-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i])
